拆分astronomy
